+++ registry/registry.go
@@ -12,7 +12,7 @@ var (
 
 // Registry is a store of image metadata.
 type Registry interface {
-	GetRepositoryImages(image.Name) ([]image.Info, error)
+	GetImageRepositoryMetadata(image.Name) (image.RepositoryMetadata, error)
 	GetImage(image.Ref) (image.Info, error)
 }